requirements.txt
psutil==6.1.0
ceo-py>=0.6.0rc0
python-dotenv==1.0.1
sympy==1.13.3
setuptools>=73.0.1
colorama>=0.4.6
duckduckgo_search>=6.3.3